rename Loan subpackage
